strutil: move out test_colored_string.py
